// pulso-propiedades-intranet/src/services/properties/types/property.type.ts
export type Property = {
  id?: string
  propertyTitle?: string
  propertyDescription?: string
  typeOfOperationId?: string
  typeOfPropertyId?: string
  timeAvailableStart?: Date | null
  timeAvailableEnd?: Date | null
  currencyId?: string
  propertyPrice?: number
  highlighted?: boolean
  isActive?: boolean
  observations?: string
  disabledReason: {
    id: string
    name: string
  }
  user?: {
    id?: string
    name?: string
    lastName?: string
  }
  customer?: {
    id?: string
    name?: string
    lastName?: string
  }
  externalLink: string
  characteristics?: {
    surface?: string
    constructedSurface?: string
    surfaceUnit?: string
    floors?: string
    numberOfFloors?: string
    terraces?: string
    bedrooms?: string
    bathrooms?: string
    typeOfKitchen?: string
    hasHeating?: boolean
    hasKitchen?: boolean
    typeOfHeating?: string
    hasAirConditioning?: boolean
    hasParking?: boolean
    hasGarage?: boolean
    numberOfParkingSpaces?: string
    hasElevator?: boolean
    hasGym?: boolean
    hasSwimmingPool?: boolean
    hasSecurity?: boolean
    typeOfSecurity?: string[] // O puedes especificar un tipo más específico aquí
    locatedInCondominium?: boolean
    isFurnished?: boolean
    hasBarbecueArea?: boolean
    numberOfPrivate?: string
    numberOfVacantFloors?: string
    numberOfMeetingRooms?: string
    hasKitchenet?: boolean
    hasHouse?: boolean
    locatedInGallery?: boolean
    locatedFacingTheStreet?: boolean
    floorLevelLocation?: string
    officeNumber?: string
    commonExpenses?: string
  }
  address?: {
    address?: string
    countryId: number
    stateId: number
    cityId: number
    letter?: string
    number?: string
    references?: string
    addressPublic?: string
  }
  images?: any
  createdAt?: Date | string
  updatedAt?: Date | string
  deletedAt?: Date | string
  property_portales?: { id: string; name: string }[]
}